Fixed Crash on Exit Bug Created in Prev Commit
The code to delete auto-backup files (see prev commit) was failing to make sure there was a project open before deleting the files.

+++ src/main/java/edu/umich/soar/visualsoar/MainFrame.java
@@ -319,14 +319,15 @@ public class MainFrame extends JFrame implements Kernel.StringEventInterface
 		}
 
 		//Check for auto-backups of the project files
-		OperatorRootNode orn = (OperatorRootNode)(operatorWindow.getModel().getRoot());
-		File projectBackupFile = new File(orn.getProjectFile() + "~");
-		if (projectBackupFile.exists()) projectBackupFile.delete();
-		File dataMapBackupFile = new File(orn.getDataMapFile() + "~");
-		if (dataMapBackupFile.exists()) dataMapBackupFile.delete();
-		File commentBackupFile = new File(dataMapBackupFile.getParent() + File.separator + "comment.dm~");
-		if (commentBackupFile.exists()) commentBackupFile.delete();
-
+		if (operatorWindow != null) {
+			OperatorRootNode orn = (OperatorRootNode) (operatorWindow.getModel().getRoot());
+			File projectBackupFile = new File(orn.getProjectFile() + "~");
+			if (projectBackupFile.exists()) projectBackupFile.delete();
+			File dataMapBackupFile = new File(orn.getDataMapFile() + "~");
+			if (dataMapBackupFile.exists()) dataMapBackupFile.delete();
+			File commentBackupFile = new File(dataMapBackupFile.getParent() + File.separator + "comment.dm~");
+			if (commentBackupFile.exists()) commentBackupFile.delete();
+		}
 	}
 
 	/**
